--- src/main/java/nextstep/line/apllication/LineService.java
@@ -1,12 +1,0 @@
-package nextstep.line.apllication;
-
-import nextstep.line.dto.LineRequest;
-import nextstep.line.dto.LineResponse;
-import org.springframework.stereotype.Service;
-
-@Service
-public class LineService {
-    public LineResponse create(LineRequest lineRequest) {
-        return null;
-    }
-}
