+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -1,4 +1,6 @@
-package nextstep.line.dto;
+package nextstep.subway.line.dto;
+
+import nextstep.subway.line.domain.Line;
 
 public class LineRequest {
     private String name;
@@ -9,6 +11,10 @@ public class LineRequest {
 
     public static LineRequest of(String name) {
         return new LineRequest(name);
+    }
+
+    public Line toLine() {
+        return new Line(name);
     }
 
     public String getName() {
