+++ src/main/java/nextstep/subway/line/ui/LineController.java
@@ -1,7 +1,7 @@
-package nextstep.line.ui;
+package nextstep.subway.line.ui;
 
-import nextstep.line.dto.LineRequest;
-import nextstep.line.dto.LineResponse;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
